Move docker helper function into appropriate mod
