+++ crates/module-system/hyperlane/tests/integration/with_agent/helpers/docker.rs
@@ -1,4 +1,5 @@
 use std::time::Duration;
+use testcontainers::core::client::docker_client_instance;
 use testcontainers::core::ExecResult;
 use testcontainers::ContainerAsync;
 use tokio::io::AsyncBufReadExt;
@@ -41,4 +42,25 @@ pub async fn print_logs_from_exec_result(name: &str, result: &mut ExecResult, ti
     })
     .await;
     tracing::info!("---- Done printing stderr");
+}
+
+#[cfg_attr(target_os = "macos", allow(dead_code))]
+pub async fn get_docker_gateway_ip() -> String {
+    let bridge_info = docker_client_instance()
+        .await
+        .unwrap()
+        .inspect_network(
+            "bridge",
+            None::<testcontainers::bollard::query_parameters::InspectNetworkOptions>,
+        )
+        .await
+        .unwrap();
+    bridge_info
+        .ipam
+        .expect("no IPAM driver found")
+        .config
+        .expect("IPAM has no configuration")
+        .into_iter()
+        .find_map(|conf| conf.gateway)
+        .expect("No gateway config in IPAM")
 }
